+++ src/main/java/com/webank/webase/front/abi/AbiService.java
@@ -130,7 +130,7 @@ public class AbiService {
 
     public void delete(Long id) {
         checkAbiIdExist(id);
-        abiRepository.delete(id);
+        abiRepository.deleteById(id);
     }
 
     public AbiInfo getAbiById(Long abiId) {