+++ src/main/java/com/webank/webase/front/event/EventService.java
@@ -295,7 +295,7 @@ public class EventService {
                                                       String queueName) {
         log.debug("unregisterNewBlock appId:{},groupId:{},exchangeName:{},queueName:{}",
                 appId, groupId, exchangeName, queueName);
-        NewBlockEventInfo eventInfo = newBlockEventInfoRepository.findById(infoId);
+        NewBlockEventInfo eventInfo = newBlockEventInfoRepository.findById(infoId).orElse(null);
         if (Objects.isNull(eventInfo)) {
             throw new FrontException(ConstantCode.DATA_NOT_EXIST_ERROR);
         }
@@ -310,7 +310,7 @@ public class EventService {
             throw new FrontException(ConstantCode.UNREGISTER_FAILED_ERROR);
         }
         // remove from db
-        newBlockEventInfoRepository.delete(infoId);
+        newBlockEventInfoRepository.deleteById(infoId);
         return newBlockEventInfoRepository.findByAppId(appId);
     }
 
@@ -347,7 +347,7 @@ public class EventService {
                                                       String queueName) {
         log.debug("unregisterContractEvent infoId:{},appId:{},groupId:{},exchangeName:{},queueName:{}",
                 infoId, appId, groupId, exchangeName, queueName);
-        ContractEventInfo eventInfo = contractEventInfoRepository.findById(infoId);
+        ContractEventInfo eventInfo = contractEventInfoRepository.findById(infoId).orElse(null);
         if (Objects.isNull(eventInfo)) {
             throw new FrontException(ConstantCode.DATA_NOT_EXIST_ERROR);
         }
@@ -367,7 +367,7 @@ public class EventService {
             throw new FrontException(ConstantCode.UNREGISTER_FAILED_ERROR);
         }
         // remove from db
-        contractEventInfoRepository.delete(infoId);
+        contractEventInfoRepository.deleteById(infoId);
         return contractEventInfoRepository.findByAppId(appId);
     }
 
@@ -385,7 +385,7 @@ public class EventService {
         log.info("getContractEventLog eventParam:{}", eventParam);
         // final CompletableFuture<List<EventLog>> callbackFuture = new CompletableFuture<>();
         final CompletableFuture<List<DecodedEventLog>> callbackFuture = new CompletableFuture<>();
-        ABICodec abiCodec = new ABICodec(cryptoSuite);
+        ABICodec abiCodec = new ABICodec(cryptoSuite, true);
         SyncEventLogCallback callback = new SyncEventLogCallback(abiCodec, abi,
             eventTopicParam.getEventName().split("\\(")[0], callbackFuture);
         EventSubscribe eventSubscribe = bcosSDK.getEventSubscribe(groupId);